Refactor: reorganize Dropzone modules
* Separate Styled components from main components
* Separate Dropzone Inner into its own folder
