+++ src/components/base/Dropzone/DropzoneInner/DropzoneInner.style.tsx
@@ -1,0 +1,8 @@
+import { styled } from '@mui/material';
+
+const StyledDropzoneInner = styled('div')({
+  pointerEvents: 'none',
+  userSelect: 'none',
+});
+
+export default StyledDropzoneInner;
