+++ src/components/base/Dropzone/DropzoneInner/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DropzoneInner';
